Moved helper function to utils.
